Fix : 스웨거 선거관련 API 분리
