+++ lawmaking/src/main/java/com/everyones/lawmaking/common/dto/ProportionalPartyImageListDto.java
@@ -4,11 +4,13 @@ import com.everyones.lawmaking.domain.entity.Party;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
 @Builder
 @Getter
+@AllArgsConstructor
 @JsonNaming(value = PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class ProportionalPartyImageListDto {
