+++ lawmaking/src/main/java/com/everyones/lawmaking/common/dto/response/ProportionalPartyImageListResponse.java
@@ -3,6 +3,7 @@ package com.everyones.lawmaking.common.dto.response;
 import com.everyones.lawmaking.common.dto.ProportionalPartyImageListDto;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -10,6 +11,7 @@ import java.util.List;
 
 @Getter
 @Builder
+@AllArgsConstructor
 @JsonNaming(value = PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class ProportionalPartyImageListResponse {
